requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
anthropic==0.25.7
    # via langchain-anthropic
anyio==4.3.0
    # via
    #   anthropic
    #   httpx
appdirs==1.4.4
    # via yfinance
apscheduler==3.6.3
    # via python-telegram-bot
argcomplete==3.3.0
    # via zappa
asgiref==3.7.2
    # via django
asttokens==2.4.1
    # via stack-data
attrs==23.2.0
    # via aiohttp
awslambdaric==2.0.11
    # via -r requirements.in
bcrypt==4.1.2
    # via paramiko
beautifulsoup4==4.12.3
    # via
    #   -r requirements.in
    #   yfinance
boto3==1.34.97
    # via
    #   kappa
    #   zappa
botocore==1.34.97
    # via
    #   boto3
    #   s3transfer
build==1.0.3
    # via pip-tools
cachetools==4.2.2
    # via python-telegram-bot
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   python-telegram-bot
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
cfn-flip==1.3.0
    # via troposphere
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   cfn-flip
    #   kappa
    #   pip-tools
cryptography==42.0.5
    # via paramiko
dataclasses-json==0.6.5
    # via
    #   langchain
    #   langchain-community
decorator==5.1.1
    # via
    #   fabric
    #   ipython
defusedxml==0.7.1
    # via langchain-anthropic
deprecated==1.2.14
    # via fabric
distro==1.9.0
    # via anthropic
django==5.0.4
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-model-utils
    #   django-s3-sqlite
    #   sentry-sdk
django-dotenv==1.4.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-model-utils==4.5.1
    # via -r requirements.in
django-s3-sqlite==0.0.5
    # via -r requirements.in
durationpy==0.6
    # via zappa
executing==2.0.1
    # via stack-data
fabric==3.2.2
    # via -r requirements.in
filelock==3.14.0
    # via huggingface-hub
frozendict==2.4.4
    # via yfinance
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.3.1
    # via huggingface-hub
h11==0.14.0
    # via httpcore
hjson==3.1.0
    # via zappa
html5lib==1.1
    # via yfinance
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via anthropic
huggingface-hub==0.23.0
    # via tokenizers
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
inflection==0.5.1
    # via nasdaq-data-link
invoke==2.2.0
    # via fabric
ipython==8.24.0
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jinja2==3.1.3
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
    #   zappa
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
kappa==0.6.0
    # via zappa
krakenex==2.2.1
    # via pykrakenapi
langchain==0.1.17
    # via -r requirements.in
langchain-anthropic==0.1.11
    # via -r requirements.in
langchain-community==0.0.36
    # via langchain
langchain-core==0.1.50
    # via
    #   langchain
    #   langchain-anthropic
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.0.1
    # via langchain
langsmith==0.1.54
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==5.2.1
    # via yfinance
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.21.2
    # via dataclasses-json
matplotlib-inline==0.1.7
    # via ipython
more-itertools==10.2.0
    # via nasdaq-data-link
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multitasking==0.0.11
    # via yfinance
mypy-extensions==1.0.0
    # via typing-inspect
nasdaq-data-link==1.0.4
    # via -r requirements.in
numpy==1.24.4
    # via
    #   langchain
    #   langchain-community
    #   nasdaq-data-link
    #   pandas
    #   yfinance
orjson==3.10.3
    # via langsmith
packaging==23.2
    # via
    #   build
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
pandas==2.0.3
    # via
    #   nasdaq-data-link
    #   pykrakenapi
    #   yfinance
paramiko==3.4.0
    # via fabric
parso==0.8.4
    # via jedi
peewee==3.17.5
    # via yfinance
pexpect==4.9.0
    # via ipython
pip-tools==7.4.0
    # via -r requirements.in
placebo==0.9.0
    # via
    #   kappa
    #   zappa
prettytable==3.10.0
    # via -r requirements.in
prompt-toolkit==3.0.43
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.22
    # via cffi
pydantic==1.10.12
    # via
    #   -r requirements.in
    #   anthropic
    #   langchain
    #   langchain-core
    #   langsmith
pygments==2.18.0
    # via ipython
pykrakenapi==0.3.2
    # via -r requirements.in
pynacl==1.5.0
    # via paramiko
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.8.2
    # via
    #   botocore
    #   nasdaq-data-link
    #   pandas
    #   zappa
python-slugify==8.0.4
    # via zappa
python-telegram-bot==13.13
    # via -r requirements.in
pytz==2024.1
    # via
    #   apscheduler
    #   pandas
    #   python-telegram-bot
    #   yfinance
pyyaml==6.0.1
    # via
    #   cfn-flip
    #   huggingface-hub
    #   kappa
    #   langchain
    #   langchain-community
    #   langchain-core
    #   zappa
requests==2.31.0
    # via
    #   huggingface-hub
    #   krakenex
    #   langchain
    #   langchain-community
    #   langsmith
    #   nasdaq-data-link
    #   yfinance
    #   zappa
s3transfer==0.10.1
    # via boto3
sentry-sdk[django]==2.1.1
    # via -r requirements.in
simplejson==3.18.4
    # via awslambdaric
six==1.16.0
    # via
    #   apscheduler
    #   asttokens
    #   cfn-flip
    #   html5lib
    #   nasdaq-data-link
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.30
    # via
    #   langchain
    #   langchain-community
sqlparse==0.4.4
    # via django
stack-data==0.6.3
    # via ipython
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
text-unidecode==1.3
    # via python-slugify
tokenizers==0.19.1
    # via anthropic
toml==0.10.2
    # via zappa
tornado==6.4
    # via python-telegram-bot
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   zappa
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
troposphere==4.8.0
    # via zappa
typing-extensions==4.11.0
    # via
    #   anthropic
    #   huggingface-hub
    #   ipython
    #   pydantic
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
tzlocal==5.2
    # via apscheduler
urllib3==2.2.1
    # via
    #   botocore
    #   requests
    #   sentry-sdk
wcwidth==0.2.13
    # via
    #   prettytable
    #   prompt-toolkit
webencodings==0.5.1
    # via html5lib
werkzeug==3.0.2
    # via zappa
wheel==0.42.0
    # via
    #   pip-tools
    #   zappa
wrapt==1.16.0
    # via deprecated
yarl==1.9.4
    # via aiohttp
yfinance==0.2.38
    # via -r requirements.in
zappa==0.58.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
